perf(macros): feature-gate the database codegen behind `db` (#2309)
`autumn-macros` had no `[features]` section at all, so `model.rs` and
`repository.rs` — together ~40k of the crate's ~60k lines — were compiled in
full for every consuming app, including one that can reach neither macro:
`autumn-web` already gates both re-exports on its own `db` feature.

Measured on a 4-core box (debug, `CARGO_INCREMENTAL=0`), cfg'ing modules out
one at a time:

  as-is                                31.2s
  minus repository.rs                   5.4s
  minus repository/model/commentable    2.5s

`repository.rs` alone is ~83% of the crate's compile time. The crate sits on
the serial critical path of a first build — nothing else starts until it
finishes — so that time lands directly on cold-start onboarding.

- `autumn-macros` gains `default = ["db"]`, with `db` gating the `model`,
  `repository`, and `commentable` modules and the two proc-macro entry points.
  Default-on keeps the full surface for a direct dependant and for
  `cargo test -p autumn-macros`.
- The serde / JSON-schema field helpers `#[derive(OpenApiSchema)]` shares with
  `#[model]` move out of `model.rs` into a new always-compiled `schema`
  module, so the derive keeps working — and keeps its tests — with `db` off.
- `autumn-web` takes the macro crate with `default-features = false` and
  forwards `db -> autumn-macros/db`. `autumn-edge` takes it
  `default-features = false` too, so enabling `edge` cannot feature-unify the
  database codegen back on for an app that does not want it.

Net: an `autumn-macros` debug build drops 31.07s -> 2.67s (-91%) with `db`
off. A database-backed app enables `db` and is unaffected — same macros, same
expansions. 921 unit tests pass with `db`, 432 with `--no-default-features`.

Refs #2309
